backend/requirements.txt
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-HTTPAuth==4.8.0
Whoosh==2.7.4
rdkit
gunicorn
python-dotenv
pillow
pandas>=1.5.0
